v5.1.7
- (Fix) CTB and GOO: Set lift values to fixed 0.05 and TSMC values to 0 for tilting vat printers (#1052)
- (Upgrade) OpenCV from 4.10.0 to 4.11.0
- (Upgrade) .NET from 9.0.7 to 9.0.8

+++ build/cvextern.ps1
@@ -14,32 +14,61 @@ $buildFile = "platforms\windows\Build_Binary_x86.bat"
 $buildArgs = '64 mini commercial no-openni no-doc no-package build'
 $customBuild = '-DWITH_EIGEN:BOOL=FALSE -DWITH_MSMF:BOOL=FALSE -DWITH_DSHOW:BOOL=FALSE -DWITH_FFMPEG:BOOL=FALSE -DWITH_GSTREAMER:BOOL=FALSE -DWITH_1394:BOOL=FALSE -DVIDEOIO_ENABLE_PLUGINS:BOOL=FALSE -DBUILD_opencv_videoio:BOOL=FALSE -DBUILD_opencv_gapi:BOOL=FALSE -DWITH_PROTOBUF:BOOL=FALSE -DBUILD_PROTOBUF:BOOL=FALSE'
 
-if ($args.Count -gt 0){
-    if($args[0] -eq 'clean'){
-        if(Test-Path -Path "$libFolder"){
-            Remove-Item -Force -Recurse -Path "$libFolder"
+$prompt = $null
+$branch = $null
+if ($args.Count -gt 0) {
+    if (-not ([string]::IsNullOrWhiteSpace($args[0]))) {
+        if($args[0] -eq 'clean') {
+            if(Test-Path -Path "$libFolder") {
+                Remove-Item -Force -Recurse -Path "$libFolder"
+            }
+            exit;
+        } else {
+            $prompt = $args[0]
         }
-        exit;
     }
 }
 
-if (-not (Test-Path -Path "$libFolder/$buildFile" -PathType Leaf)) {
-    $confirmation = Read-Host "$libFolder directory does not exists, do you want to download it?
-y/yes: Download master branch
-4.7.0: Download specific tag
+$foundCMake = [bool] (Get-Command -ErrorAction Ignore -Type Application cmake)
+if (!$foundCMake)
+{
+    Write-Error "cmake not found, please install via Visual Studio."
+	exit;
+}
+
+$foundDotnet = [bool] (Get-Command -ErrorAction Ignore -Type Application dotnet)
+if (!$foundDotnet)
+{
+    Write-Error "dotnet not found, please install."
+	exit;
+}
+
+if (-not $prompt) {
+    $prompt = Read-Host "Select a branch or tag to build cvextern.dll
+y/yes/master/main: Download master branch
+4.9.0: Download specific tag
 n/no: Cancel
 Option"
+}
 
-    if ($confirmation -eq 'y' -or $confirmation -eq 'yes') {
-        Write-Output "Clone master"
-        git clone --recurse-submodules --depth 1 "https://github.com/emgucv/emgucv" "$libFolder"
-    }elseif($confirmation -match '^[0-9]+[.][0-9]+[.][0-9]+$') {
-        Write-Output "$confirmation"
-        git clone --recurse-submodules --depth 1 --branch "$confirmation" "https://github.com/emgucv/emgucv" "$libFolder"
-    }
-    else {
-        exit;
-    }
+if ($prompt -eq 'y' -or $prompt -eq 'yes' -or $prompt -eq 'master' -or $prompt -eq 'main') {
+    $branch = 'master'
+} elseif($prompt -match '^[0-9]+[.][0-9]+[.][0-9]+$') {
+    $branch = $prompt
+} else {
+    exit;
+}
+
+if (-not $branch) {
+    Write-Error "Invalid branch or tag specified."
+    exit;
+}
+
+$libFolder = "$libFolder-$branch"
+
+if (-not (Test-Path -Path "$libFolder/$buildFile" -PathType Leaf)) {
+    Write-Output "Clonning $branch of EmguCV to $libFolder"
+    git clone --recurse-submodules --depth 1 --branch "$branch" "https://github.com/emgucv/emgucv" "$libFolder"
 }
 
 Set-Location "$libFolder"
